% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_l_function.R
\name{center_l_function}
\alias{center_l_function}
\title{center_l_function}
\usage{
center_l_function(x, ...)
}
\arguments{
\item{x}{ppp}

\item{...}{Arguments passed to \code{spatstat::Lest()}}
}
\description{
Centered L-function
}
\details{
Center Besag's L-function to zero.
}
\examples{
input_pattern <- spatstat::runifpoint(n = 100)
center_l_function(input_pattern, correction = "Ripley")

lest <- spatstat::Lest(input_pattern)
center_l_function(lest)

}
\references{
Besag, J. (1977) Discussion of Dr Ripley's paper. Journal of the Royal Statistical Society, Series B, 39, 193–195.
}
\seealso{
\code{\link{envelope}}
\code{\link{density.ppp}}
}
