% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_o_ring.R
\name{estimate_o_ring}
\alias{estimate_o_ring}
\title{estimate_o_ring}
\usage{
estimate_o_ring(x, ...)
}
\arguments{
\item{x}{ppp}

\item{...}{Arguments passed to \code{spatstat::pcf.ppp()}}
}
\description{
O-ring function
}
\details{
Estimate the O-ring function.
}
\examples{
input_pattern <- spatstat::runifpoint(n = 100)
estimate_o_ring(input_pattern)

}
\references{
Wiegand, T. and Moloney, K. A. 2014. Handbook of spatial point-pattern analysis in ecology. - Chapman and Hall/CRC Press.

Wiegand, T. and Moloney, K. A. 2004. Rings, circles, and null models for point pattern analysis in ecology. - Oikos 104: 209–229.
}
\seealso{
\code{\link{envelope}}
\code{\link{density.ppp}}
}
