% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onpoint-package.R
\docType{package}
\name{onpoint}
\alias{onpoint}
\alias{onpoint-package}
\title{onpoint}
\description{
Growing collection of small helper functions for point pattern analysis. Most functions
are designed to work with the spatstat package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatialecology.github.io/onpoint/}
  \item Report bugs at \url{https://github.com/r-spatialecology/onpoint/issues}
}

}
\author{
\strong{Maintainer}: Maximilian H.K. Hesselbarth \email{mhk.hesselbarth@gmail.com} (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})

}
\keyword{internal}
