% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.env_summarised.R
\name{plot.env_summarised}
\alias{plot.env_summarised}
\title{plot.env_summarised}
\usage{
\method{plot}{env_summarised}(
  x,
  col_poly = c("#97CBDE", "#E1B0B5"),
  x_lab = NULL,
  y_lab = NULL,
  base_size = 10,
  label = TRUE,
  ...
)
}
\arguments{
\item{x}{Random patterns.}

\item{col_poly}{Colours for areas above and below envelope.}

\item{x_lab, y_lab}{Labels of x- and y-axis.}

\item{base_size}{Base size of plot}

\item{label}{If TRUE the ratios of the area above and below are added to the plot.}

\item{...}{To be generic for plotting function.}
}
\description{
Plotting method for env_summarised object
}
\details{
Ploting method for summarised envelope created with \code{\link{summarise_envelope}}.
}
\examples{
set.seed(42)
input_pattern <- spatstat::rThomas(kappa = 15, scale = 0.05, mu = 5)

cluster_env <- spatstat::envelope(input_pattern, fun = "pcf", nsim = 39,
funargs = list(divisor = "d", correction = "Ripley", stoyan = 0.25))

x <- summarise_envelope(cluster_env)
plot(x)

}
\seealso{
\code{\link{summarise_envelope}}
}
