% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.env_summarised.R
\name{print.env_summarised}
\alias{print.env_summarised}
\title{print.env_summarised}
\usage{
\method{print}{env_summarised}(x, return_area = FALSE, digits = 2, ...)
}
\arguments{
\item{x}{Random patterns.}

\item{return_area}{If true, not the ratio but the area is returned.}

\item{digits}{Number of decimal places (round).}

\item{...}{Arguments passed to cat}
}
\description{
Print method for env_summarised object
}
\details{
Printing method for summarised envelope created with \code{\link{summarise_envelope}}.
}
\examples{
set.seed(42)
input_pattern <- spatstat::rThomas(kappa = 15, scale = 0.05, mu = 5)

cluster_env <- spatstat::envelope(input_pattern, fun = "pcf", nsim = 39,
funargs = list(divisor = "d", correction = "Ripley", stoyan = 0.25))

x <- summarise_envelope(cluster_env)
print(x)

}
\seealso{
\code{\link{summarise_envelope}}
}
