% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlabel_local.R
\name{rlabel_local}
\alias{rlabel_local}
\title{rlabel_local}
\usage{
rlabel_local(X, distance, nsim = 1, drop = TRUE, verbose = TRUE)
}
\arguments{
\item{X}{ppp}

\item{distance}{Mark of points that do not change.}

\item{nsim}{Number of patterns to simulate.}

\item{drop}{If nsim = 1 and drop = TRUE , the result will be a point pattern,
rather than a list containing a point pattern.}

\item{verbose}{If TRUE warning messages are printed.}
}
\description{
Local random labelling of marked point pattern
}
\details{
Local random labelling function, i.e. marks will be shuffeld only across points
within the specified local distance. Technically, this is achived by sampling the
mark of a neighbouring point j within the distance d for the focal point i. Thus,
the distance d must be selected in a way that each point has at least one neighbour
within d.
}
\examples{
set.seed(42)
pattern <- spatstat::runifpoint(n = 250, win = spatstat::owin(c(0, 100), c(0, 100)))
spatstat::marks(pattern) <- runif(n = 250, min = 10, max = 120)

rlabel_local(X = pattern, distance = 25, nsim = 3)

}
\references{
Velazquez, E. et al. 2016. An evaluation of the state of spatial point pattern analysis in ecology. - Ecography (Cop.). 39: 1-14.

Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in
ecology. Chapman and Hall/CRC Press, Boca Raton.
}
\seealso{
\code{\link{rlabel}}
}
