% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_heterogenous_pattern.R
\name{simulate_heterogenous_pattern}
\alias{simulate_heterogenous_pattern}
\title{simulate_heterogenous_pattern}
\usage{
simulate_heterogenous_pattern(x, nsim, ...)
}
\arguments{
\item{x}{ppp}

\item{nsim}{Number of patterns to simulate.}

\item{...}{Arguments passed to \code{spatstat::density.ppp()}}
}
\description{
Simulate heterogenous pattern
}
\details{
Simulate heterogenous point patterns as null model data for \code{spatstat::envelope()}.
A heterogenous Poisson process is used, meaning that there are no interaction between points,
however, the simulated coordinates depend on the intensity \eqn{\lambda} of the input pattern.
}
\examples{
input_pattern <- spatstat::rpoispp(lambda = function(x , y) {100 * exp(-3 * x)}, nsim = 1)
null_model <- simulate_heterogenous_pattern(input_pattern, nsim = 19)
spatstat::envelope(Y = input_pattern, fun = spatstat::pcf, nsim = 19, simulate = null_model)

}
\references{
Baddeley, A. et al. 2015. Spatial Point Patterns: Methodology and Applications with R. - Chapman and Hall/CRC Press.

Wiegand, T. and Moloney, K. A. 2014. Handbook of spatial point-pattern analysis in ecology. - Chapman and Hall/CRC Press.
}
\seealso{
\code{\link{envelope}} \cr
\code{\link{density.ppp}}
}
