% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_envelope.R
\name{summarise_envelope}
\alias{summarise_envelope}
\title{summarise_envelope}
\usage{
summarise_envelope(x)
}
\arguments{
\item{x}{fv}
}
\description{
Summarise a simulation envelope
}
\details{
Summarise a simulation envelope
}
\examples{
set.seed(42)
input_pattern <- spatstat::rThomas(kappa = 15, scale = 0.05, mu = 5)

cluster_env <- spatstat::envelope(input_pattern, fun = "pcf", nsim = 39,
funargs = list(divisor = "d", correction = "Ripley", stoyan = 0.25))

summarise_envelope(cluster_env)

}
\seealso{
\code{\link{envelope}}
}
