% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_l_function.R
\name{center_l_function}
\alias{center_l_function}
\title{center_l_function}
\usage{
center_l_function(x, ...)
}
\arguments{
\item{x}{ppp}

\item{...}{Arguments passed to \code{spatstat.core::Lest()}}
}
\description{
Centered L-function
}
\details{
Centers Besag's L-function to zero by calculating L(r) -r. Centering the L-function
allows an easier interpretation and plotting of the results (Haase 1995).
}
\examples{
input_pattern <- spatstat.core::runifpoint(n = 100)
center_l_function(input_pattern, correction = "Ripley")

lest <- spatstat.core::Lest(input_pattern)
center_l_function(lest)

}
\references{
Besag, J. E. 1977. Discussion on Dr. Ripley's paper. - J. R. Stat. Soc. Ser. B 39: 193-195.

Ripley, B. D. 1977. Modelling spatial patterns. - J. R. Stat. Soc. Ser. B 39: 172-192.

Haase, P. 1995. Spatial pattern analysis in ecology based on Ripley's K-function:
Introduction and methods of edge correction. - J. Veg. Sci. 6: 575-582.
}
\seealso{
\code{\link{Lest}}
}
