% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_l_function.R
\name{center_l_function}
\alias{center_l_function}
\title{center_l_function}
\usage{
center_l_function(x, ...)
}
\arguments{
\item{x}{ppp}

\item{...}{Arguments passed to \code{spatstat.core::Lest()}}
}
\value{
fv.object
}
\description{
Centered L-function
}
\details{
Centers Besag's L-function to zero by calculating L(r) -r. Centering the L-function
allows an easier interpretation and plotting of the results (Haase 1995).

Returns an 'Function value object' of the \code{spatstat} package.
}
\examples{
input_pattern <- spatstat.core::runifpoint(n = 100)
center_l_function(input_pattern, correction = "Ripley")

lest <- spatstat.core::Lest(input_pattern)
center_l_function(lest)

}
\references{
Besag, J.E., 1977. Discussion on Dr. Ripley’s paper. Journal of the Royal
Statistical Society. Series B (Methodological) 39, 193–195.
https://doi.org/10.1111/j.2517-6161.1977.tb01616.x

Ripley, B.D., 1977. Modelling spatial patterns. Journal of the Royal Statistical
Society. Series B (Methodological) 39, 172–192.
https://doi.org/10.1111/j.2517-6161.1977.tb01615.x

Haase, P., 1995. Spatial pattern analysis in ecology based on Ripley’s K-function:
Introduction and methods of edge correction. Journal of Vegetation Science 6, 575–582.
https://doi.org/10.2307/3236356
}
\seealso{
\code{\link{Lest}}
}
