% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantum_plot.R
\name{quantum_plot}
\alias{quantum_plot}
\title{quantum_plot}
\usage{
quantum_plot(input, labels = c("clustering", "randomness",
  "segregation"), title = NULL, xlab = NULL, ylab = NULL, size = 5,
  full_fun = T, standarized = F)
}
\arguments{
\item{input}{envelope.}

\item{labels}{Name of the labels. See details for more information.}

\item{title}{Plot title.}

\item{xlab, ylab}{axis labels.}

\item{size}{Size of the colour bar.}

\item{full_fun}{If true observed value and envelope is plotted.}

\item{standarized}{If true observed value is standardized. See details for more details.}
}
\description{
Plot simulation envelopes
}
\details{
This functions provides a plotting style for envelope objects of the spatstat
package (for more information please see `?spatstat::envelope`). The location of the
observed value in relation to the simulation envelope of the null model data is
indicated by an additional colour bar at the bottom of the plot.

A named vector including labels for the following three cases:
\cr 1 = obs > hi
\cr 2 = lo < obs < hi
\cr 3 = obs < lo

All values are standarized by subtracting the theoretical value for CSR
}
\examples{
set.seed(42)
pattern <- spatstat::rThomas(kappa = 50, scale = 0.025, mu = 5)
csr_envelope <- spatstat::envelope(pattern, fun = spatstat::pcf, nsim = 39)
quantum_plot(csr_envelope, ylab = "g(r)")

}
\references{
Esser D.S., Leveau J.H.J., Meyer K.M., Wiegand K. (2015): Spatial scales of
interactions among bacteria and between bacteria and the leaf surface. FEMS Microbiology Ecology, in press.
}
\seealso{
\code{\link{envelope}}
}
