% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_heterogenous_pattern.R
\name{simulate_heterogenous_pattern}
\alias{simulate_heterogenous_pattern}
\title{simulate_heterogenous_pattern}
\usage{
simulate_heterogenous_pattern(x, nsim, ...)
}
\arguments{
\item{x}{ppp}

\item{nsim}{Number of patterns to simulate.}

\item{...}{Arguments passed to \code{spatstat::density.ppp()}}
}
\description{
Simulate heterogenous pattern
}
\details{
Simulate heterogenous point patterns as null model data for \code{spatstat::envelope()}.
}
\examples{
input_pattern <- spatstat::rpoispp(lambda = function(x , y) {100 * exp(-3 * x)}, nsim = 1)
null_model <- simulate_heterogenous_pattern(input_pattern, nsim = 19)
spatstat::envelope(Y = input_pattern, fun = spatstat::pcf, nsim = 19, simulate = null_model)

}
\seealso{
\code{\link{envelope}}
\code{\link{density.ppp}}
}
