% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at.R
\name{split_at}
\alias{split_at}
\title{split_at}
\usage{
split_at(x, pos)
}
\arguments{
\item{x}{vector with positions to split.}
}
\description{
Split vector
}
\details{
Split vector at position(s).
}
\keyword{internal}
