% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_envelope.R
\name{summarise_envelope}
\alias{summarise_envelope}
\title{summarise_envelope}
\usage{
summarise_envelope(x, seperated = TRUE, plot_result = FALSE)
}
\arguments{
\item{x}{fv}

\item{seperated}{If TRUE one value for the relative area above and one for
the relative area below the envelope are returned.}

\item{plot_result}{A plot is drawn.}
}
\value{
vector
}
\description{
Summarise a simulation envelope
}
\details{
Summarise a simulation envelope. The area above and below the null model
envelope is divided by the area where the observed value is within the envelope.
If \code{seperated = TRUE}, the first returning value is the relative area above,
the second value the relative value below the envelope.
}
\examples{
set.seed(42)
input_pattern <- spatstat::rThomas(kappa = 15, scale = 0.05, mu = 5)

cluster_env <- spatstat::envelope(input_pattern, fun = "pcf", nsim = 39,
funargs = list(divisor = "d", correction = "Ripley", stoyan = 0.25))

summarise_envelope(cluster_env, seperated = FALSE, plot = TRUE)

}
\seealso{
\code{\link{envelope}}
}
